refactor message category parser
